worker/requirements.txt
pika
requests
google-generativeai
redis[hiredis]
python-dotenv
celery
fastapi[standard]
uvicorn[standard]
boto3